Changed to have an as-is bitdepth setting along with fixing issues when trying to manually override this Fixed the sizing of some of the controls on the preferences to avoid being clipped Removed some more of the spdif related code that's not needed within this build Also added 88.2 kHz as a samplerate conversion value to match what's being used Also done bit more to reduce the size of the compiled plug-in by int -> shorts

+++ valib/valib/filters/delay.cpp
@@ -70,7 +70,7 @@ Delay::init(Speakers _spk)
 {
   int ch;
   const int nch = _spk.nch();
-  const int *order = _spk.order();
+  const short int *order = _spk.order();
   const double factor = units2samples(units);
 
   memset(ch_delays, 0, sizeof(ch_delays));
